allow metric and figure parameterization (#71)
