+++ spd/configs.py
@@ -1,6 +1,10 @@
 """Config classes of various types"""
 
-from typing import Any, ClassVar, Literal, Self
+import importlib
+import inspect
+from abc import ABC, abstractmethod
+from collections.abc import Callable
+from typing import Any, ClassVar, Literal, Self, override
 
 from pydantic import (
     BaseModel,
@@ -16,6 +20,64 @@ from pydantic import (
 from spd.log import logger
 from spd.models.components import GateType
 from spd.spd_types import ModelPath, Probability
+
+
+class _FnConfig(BaseModel, ABC):  # pyright: ignore[reportUnsafeMultipleInheritance]
+    model_config: ClassVar[ConfigDict] = ConfigDict(extra="forbid", frozen=True)
+    name: str = Field(
+        ...,
+        description="Name of the function to call",
+    )
+    extra_kwargs: dict[str, Any] = Field(
+        default={},
+        description="Keyword arguments to pass to the function",
+    )
+
+    @abstractmethod
+    def get_real_func(self) -> Callable[..., Any]: ...
+
+    @model_validator(mode="after")
+    def validate_fn_kwargs(self) -> Self:
+        real_fn = self.get_real_func()
+
+        # get its signature and drop the first 'inputs' parameter
+        sig = inspect.signature(real_fn)
+        params_after_inputs = list(sig.parameters.values())[1:]
+        sig_extra_only = inspect.Signature(params_after_inputs)
+
+        # see if our kwargs are valid
+        try:
+            sig_extra_only.bind(**self.extra_kwargs)
+        except TypeError as e:
+            # replace the error as e will include something like
+            # "unexpected parameter 'foo'" or "missing a required argument: 'bar'"
+            raise ValueError(f"Invalid kwargs for {self.name!r}: {e}") from None
+
+        return self
+
+
+class FiguresFnConfig(_FnConfig):
+    @override
+    def get_real_func(self) -> Callable[..., Any]:
+        available_funcs = importlib.import_module("spd.figures").FIGURES_FNS
+        real_fn = available_funcs.get(self.name)
+        if real_fn is None:
+            raise ValueError(
+                f"Figure function {self.name!r} not found. Available functions: {available_funcs.keys()}"
+            )
+        return real_fn
+
+
+class MetricsFnConfig(_FnConfig):
+    @override
+    def get_real_func(self) -> Callable[..., Any]:
+        available_funcs = importlib.import_module("spd.metrics").METRICS_FNS
+        real_fn = available_funcs.get(self.name)
+        if real_fn is None:
+            raise ValueError(
+                f"Metric function {self.name!r} not found. Available functions: {available_funcs.keys()}"
+            )
+        return real_fn
 
 
 class TMSTaskConfig(BaseModel):
@@ -216,15 +278,11 @@ class Config(BaseModel):
         description="Interval (in steps) at which to save model checkpoints (None disables saving "
         "until the end of training).",
     )
-    log_ce_losses: bool = Field(
-        default=False,
-        description="If True, additionally track cross-entropy losses during training",
-    )
-    metrics_fns: list[str] = Field(
+    metrics_fns: list[MetricsFnConfig] = Field(
         default=[],
         description="List of local names of functions to use for computing metrics. These functions must be defined in the `spd.metrics_and_figs` module.",
     )
-    figures_fns: list[str] = Field(
+    figures_fns: list[FiguresFnConfig] = Field(
         default=[],
         description="List of local names of functions to use for creating figures. These functions must be defined in the `spd.metrics_and_figs` module.",
     )
