+++ spd/experiments/lm/ss_emb_config.yaml
@@ -40,8 +40,14 @@ image_freq: 2000
 image_on_first_step: true
 print_freq: 1000
 save_freq: null
-figures_fns: ["ci_histograms", "mean_component_activation_counts"]
-metrics_fns: ["ci_l0", "lm_kl", "lm_ce_losses", "lm_embed"]
+figures_fns:
+  - name: "ci_histograms"
+  - name: "mean_component_activation_counts"
+metrics_fns:
+  - name: "ci_l0"
+  - name: "lm_kl"
+  - name: "lm_ce_losses"
+  - name: "lm_embed"
 
 # --- Pretrained model info ---
 pretrained_model_class: transformers.LlamaForCausalLM
