+++ spd/experiments/lm/ts_config.yaml
@@ -43,8 +43,13 @@ image_freq: 2000
 image_on_first_step: true
 print_freq: 1000
 save_freq: null
-figures_fns: ["ci_histograms", "mean_component_activation_counts"]
-metrics_fns: ["ci_l0", "lm_kl", "lm_ce_losses"]
+figures_fns:
+  - name: "ci_histograms"
+  - name: "mean_component_activation_counts"
+metrics_fns:
+  - name: "ci_l0"
+  - name: "lm_kl"
+  - name: "lm_ce_losses"
 
 
 # --- Pretrained model info ---
