+++ spd/experiments/resid_mlp/resid_mlp3_config.yaml
@@ -39,8 +39,12 @@ image_freq: 5_000
 image_on_first_step: true
 print_freq: 500
 save_freq: null
-figures_fns: ["ci_histograms", "mean_component_activation_counts", "uv_and_identity_ci"]
-metrics_fns: ["ci_l0"]
+figures_fns:
+  - name: "ci_histograms"
+  - name: "mean_component_activation_counts"
+  - name: "uv_and_identity_ci"
+metrics_fns:
+  - name: "ci_l0"
 
 # --- Pretrained model info ---
 pretrained_model_class: "spd.experiments.resid_mlp.models.ResidualMLP"
